// frontend/src/types/tourism/blog/v1/blog.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.7
//   protoc               unknown
// source: tourism/blog/v1/blog.proto

/* eslint-disable */

export const protobufPackage = "tourism.blog.v1";

export interface Comment {
  authorId: string;
  text: string;
  createdAt: string;
  lastModifiedAt: string;
  authorUsername: string;
}

export interface Blog {
  id: string;
  authorId: string;
  title: string;
  description: string;
  creationDate: string;
  images: string[];
  comments: Comment[];
  likedByUserIds: string[];
  canRead: boolean;
  authorUsername: string;
}

export interface CreateBlogRequest {
  authorId: string;
  title: string;
  description: string;
  images: string[];
}

export interface GetAllBlogsRequest {
}

export interface GetAllBlogsResponse {
  blogs: Blog[];
}

export interface GetBlogRequest {
  id: string;
}

export interface AddCommentRequest {
  userId: string;
  blogId: string;
  comment: Comment | undefined;
}

export interface ToggleLikeRequest {
  blogId: string;
  userId: string;
}